KC-25 finished up typing test to two lines
Major changes in how the typing test is ran, almost all of the logic has been moved out off app.js and into typingtest.js. This allows for cleaner code and for the typing component to be able to re-render itself with all state variable changes. Still messy, but gets the job done for now