Kc 88 multiplayer lobby mvp (#18)
* Adding client based server connection

* Update mp prototype to support multiple lobbies, separate Lobby code from Multiplayer.js

This commit introduces multiple lobby support to the multiplayer prototype and separates the Lobby functionality from the Multiplayer.js file.

* Add custom lobby button and modal

This update includes the addition of the Join Custom Lobby button. Upon clicking this button, a Modal component that contains an input box loads. The user is prompted to enter a lobby ID which then updates the lobbyID state that's used to create the lobby socket room.

* Remove unused code

* Modify multiplayer server code to run locally, implement custom lobbies

This commit involves the addition of custom lobby creation and joining. It also involves an update of the socket.io server code that allows it to run locally

* Update .gitignore to ignore package-lock.json

* Revert "Update .gitignore to ignore package-lock.json"

This reverts commit fbc30e660c650793cf612e7bf476f6ac38a0ef6e.

* Updating committed package-lock and package

Co-authored-by: ColinHarker <colinharker5@yahoo.com>