Kc 96 google account login (#21)
* KC-45 Initial commit of SocketIO prototype

Prototype is a simple chat app using a local instance of a NodeJS server and SocketIO. In this commit, the app sends a message on Submit that doesn't reflect what the user inputted.

* Revert "KC-45 Initial commit of SocketIO prototype"

This reverts commit 32bc7c82a0f1975426b88f9672b54a75f0ef3732.

* GoogleLogin button inital

This is the initial integration of google button. Does not work.

* Update SignInModal.js

* Added patchwork code so that google response can register account with DB

This comes along with bugs such that if an account is already registered this doesn't work. Will have to be reworked when Database is updated, but for now it proves functionality

* Signup and login works

Signup and login works

Co-authored-by: tuc56407 <tuc56407@temple.edu>
Co-authored-by: unknown <colinharker5@yahoo.com>