Training works with custom word bank
Removed previous concept of directly importing a text file word for word in the same order, replaced it with a word bank. You can now import a text file with words separated by spaces(multiple spaces in a row will be replaced with a single space, including next line), and all the words will be added into a word bank array. When word bank is not empty,  custom word bank will be used instead of default for generating random words.